+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/test/util/TestSetupUtils.java
@@ -2,16 +2,15 @@ package ru.vyarus.dropwizard.guice.test.util;
 
 import com.google.common.base.Stopwatch;
 import org.junit.jupiter.api.extension.ExtensionContext;
+import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook;
 import ru.vyarus.dropwizard.guice.module.installer.util.InstanceUtils;
-import ru.vyarus.dropwizard.guice.test.jupiter.env.EnableSetup;
 import ru.vyarus.dropwizard.guice.test.jupiter.env.ListenersSupport;
 import ru.vyarus.dropwizard.guice.test.jupiter.env.TestEnvironmentSetup;
+import ru.vyarus.dropwizard.guice.test.jupiter.env.TestExecutionListener;
 import ru.vyarus.dropwizard.guice.test.jupiter.env.TestExtension;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.conf.ExtensionConfig;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.conf.GuiceyTestTime;
 
-import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -63,7 +62,7 @@ public final class TestSetupUtils {
             for (TestEnvironmentSetup support : config.extensions) {
                 // required to recognize hooks registered from setup objects
                 config.tracker.setContextHook(support.getClass());
-                final Object res = support.setup(builder);
+                final Object res = setup(support, config, builder);
                 // method could return anything, but only closable object would be tracked
                 if (res instanceof AutoCloseable || res instanceof ExtensionContext.Store.CloseableResource) {
                     // would be closed automatically with storage closing
@@ -74,31 +73,20 @@ public final class TestSetupUtils {
         }
     }
 
-    /**
-     * Validate fields annotated with {@link ru.vyarus.dropwizard.guice.test.jupiter.env.EnableSetup}
-     * for correctness.
-     *
-     * @param fields                fields to validate
-     * @param includeInstanceFields true to allow instance fields, false to break if instance field detected
-     */
-    public static void validateFields(final List<Field> fields, final boolean includeInstanceFields) {
-        for (Field field : fields) {
-            if (!TestEnvironmentSetup.class.isAssignableFrom(field.getType())) {
-                throw new IllegalStateException(String.format(
-                        "Field %s annotated with @%s, but its type is not %s",
-                        toString(field), EnableSetup.class.getSimpleName(),
-                        TestEnvironmentSetup.class.getSimpleName()
-                ));
-            }
-            if (!includeInstanceFields && !Modifier.isStatic(field.getModifiers())) {
-                throw new IllegalStateException(String.format("Field %s annotated with @%s must be static",
-                        toString(field), EnableSetup.class.getSimpleName()));
-            }
+    private static Object setup(final TestEnvironmentSetup support,
+                                final ExtensionConfig config,
+                                final TestExtension builder) {
+        final Object res = support.setup(builder);
+        // auto registration of hook (note: hook still might be registered from @EnableHook annotation in the same
+        // class, but only one hook instance would be used - only two records would be in debug)
+        if (support instanceof GuiceyConfigurationHook && !config.hooks.contains(support)) {
+            builder.hooks((GuiceyConfigurationHook) support);
         }
-    }
-
-    private static String toString(final Field field) {
-        return field.getDeclaringClass().getName() + "." + field.getName();
+        // auto registration of listener
+        if (support instanceof TestExecutionListener) {
+            builder.listen((TestExecutionListener) support);
+        }
+        return res;
     }
 
     /**